pyproject.toml
[project]
name = "bgnd"
version = "0.1.0"
description = "Dynamic piggy bank virtual background for Microsoft Teams"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.9",
    "mediapipe>=0.10.9",
    "pyvirtualcam>=0.15",
    "numpy>=1.26",
    "Pillow>=10.2",
]

[tool.setuptools.packages.find]
include = ["bgnd*"]

[project.scripts]
bgnd = "bgnd.__main__:main"
